requirements.txt

joblib==1.3.2
keras==2.14.0
Markdown==3.5
matplotlib==3.8.0
ml-dtypes==0.2.0
numpy
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas
Pillow
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy
six==1.16.0
sqlparse==0.4.4
tensorboard
tensorboard-data-server
tensorflow-cpu
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug
whitenoise
wrapt==1.14.1
python-decouple
mlflow
tensorflow_model_optimization
zipfile